API

Types defined in the package

RobustModels.AbstractRobustModel — Type

AbstractRobustModel

Abstract type for robust models.

RobustModels.jl implements one subtype: RobustLinearModel. See the documentation for each for more details.

RobustModels.AbstractEstimator — Type

General location or scale estimator

RobustModels.AbstractQuantileEstimator — Type

Generalized M-Quantile estimator

RobustModels.LossFunction — Type

An estimator needs a cost/loss function for the modified (weighted) least squares problems of the form:

\[\min \sum_i \rho\left(\dfrac{r_i}{\hat{\sigma}} ight)\]

RobustModels.RobustLinearModel — Type

RobustLinearModel

Robust linear model representation

Fields

resp: the RobustLinResp structure.

pred: the predictor structure, of type DensePredChol, SparsePredChol, DensePredCG, SparsePredCG or RidgePred.

fitdispersion: if true, the dispersion is estimated otherwise it is kept fixed

fitted: if true, the model was already fitted

RobustModels.DensePredCG — Type

DensePredCG

A LinPred type with Conjugate Gradient and a dense X

Members

X: Model matrix of size n × p with n ≥ p. Should be full column rank.

beta0: base coefficient vector of length p

delbeta: increment to coefficient vector, also of length p

scratchbeta: scratch vector of length p, used in linpred! method

RobustModels.SparsePredCG — Type

SparsePredCG

A LinPred type with Conjugate Gradient and a sparse X

Members

X: Model matrix of size n × p with n ≥ p. Should be full column rank.

beta0: base coefficient vector of length p

delbeta: increment to coefficient vector, also of length p

scratchbeta: scratch vector of length p, used in linpred! method

GLM.DensePredChol — Type

DensePredChol{T}

A LinPred type with a dense Cholesky factorization of X'X

Members

X: model matrix of size n × p with n ≥ p. Should be full column rank.

beta0: base coefficient vector of length p

delbeta: increment to coefficient vector, also of length p

scratchbeta: scratch vector of length p, used in linpred! method

chol: a Cholesky object created from X'X, possibly using row weights.

scratchm1: scratch Matrix{T} of the same size as X

scratchm2: scratch Matrix{T} os the same size as X'X

GLM.SparsePredChol — Type

SparsePredChol{T<:BlasReal} <: LinPred

A LinPred type with a sparse Cholesky factorization of X'X

Members

X: model matrix of size n×p with n ≥ p. Should be full column rank.

beta0: base coefficient vector of length p

delbeta: increment to coefficient vector, also of length p

scratchbeta: scratch vector of length p, used in linpred! method

chol: a Cholesky object created from X'X, possibly using row weights.

RobustModels.RidgePred — Type

RidgePred

Regularized predictor using ridge regression on the p features.

Members

X: model matrix

λ: shrinkage parameter of the regularizer

G: regularizer matrix of size p×p.

βprior: regularizer prior of the coefficient values. Default to zeros(p).

pred: the non-regularized predictor using an extended model matrix.

pivot: for DensePredChol, if the decomposition was pivoted.

scratchbeta: scratch vector of length p, used in linpred! method

RobustModels.RobustLinResp — Type

RobustLinResp

Robust linear response structure.

Solve the following minimization problem:

\[\min \sum_i \rho\left(\dfrac{r_i}{\hat{\sigma}} ight)\]

Fields

est: estimator used for the model

y: response vector

μ: mean response vector

offset: offset added to Xβ to form μ. Can be of length 0

wts: prior case weights. Can be of length 0.

σ: current estimate of the scale or dispersion

devresid: the deviance residuals

wrkwt: working case weights for the Iteratively Reweighted Least Squares (IRLS) algorithm

wrkres: working residuals for IRLS

wrkscaledres: scaled residuals for IRLS

RobustModels.QuantileRegression — Type

QuantileRegression

Quantile regression representation

Fields

τ: the quantile value

X: the model matrix

β: the coefficients

y: the response vector

wts: the weights

wrkres: the working residuals

fitdispersion: if true, the dispersion is estimated otherwise it is kept fixed

fitted: if true, the model was already fitted

Constructors for models

Missing docstring.

Missing docstring for fit(::Type{M}, ::Union{AbstractMatrix{T},SparseMatrixCSC{T}}, ::AbstractVector{T}, ::AbstractEstimator) where {T<:AbstractFloat, M<:RobustLinearModel}. Check Documenter's build log for details.

StatsAPI.fit — Method

fit(::Type{M}, X::Union{AbstractMatrix{T},SparseMatrixCSC{T}},
    y::AbstractVector{T}; quantile::AbstractFloat=0.5,
    dofit::Bool          = true,
    wts::FPVector        = similar(y, 0),
    fitdispersion::Bool  = false,
    fitargs...) where {M<:QuantileRegression, T<:AbstractFloat}

Fit a quantile regression model with the model matrix (or formula) X and response vector (or dataframe) y.

It is solved using the exact interior method.

Arguments

X: the model matrix (it can be dense or sparse) or a formula

y: the response vector or a dataframe.

Keywords

quantile::AbstractFloat=0.5: the quantile value for the regression, between 0 and 1.

dofit::Bool = true: if false, return the model object without fitting;

wts::Vector = []: a weight vector, should be empty if no weights are used;

fitdispersion::Bool = false: reevaluate the dispersion;

fitargs...: other keyword arguments like verbose to print iteration details.

Output

the RobustLinearModel object.

RobustModels.rlm — Function

rlm(X, y, args...; kwargs...)

An alias for fit(RobustLinearModel, X, y, est; kwargs...).

The arguments X and y can be a Matrix and a Vector or a Formula and a DataFrame.

RobustModels.quantreg — Function

quantreg(X, y, args...; kwargs...)

An alias for fit(QuantileRegression, X, y; kwargs...).

The arguments X and y can be a Matrix and a Vector or a Formula and a DataFrame.

StatsAPI.fit! — Function

fit!(m::RobustLinearModel; initial_scale::Union{Symbol, Real}=:mad,
          σ0::Union{Nothing, Symbol, Real}=initial_scale,
          initial_coef::AbstractVector=[],
          β0::AbstractVector=initial_coef,
          correct_leverage::Bool=false, kwargs...)

Optimize the objective of a RobustLinearModel. When verbose is true the values of the objective and the parameters are printed on stdout at each iteration.

This function assumes that m was correctly initialized.

This function returns early if the model was already fitted, instead call refit!.

fit!(m::QuantileRegression;
     verbose::Bool=false,
     quantile::Union{Nothing, AbstractFloat}=nothing,
     correct_leverage::Bool=false,
     kwargs...)

Optimize the objective of a QuantileRegression. When verbose is true the values of the objective and the parameters are printed on stdout at each function evaluation.

This function assumes that m was correctly initialized. This function returns early if the model was already fitted, instead call refit!.

RobustModels.refit! — Function

refit!(m::RobustLinearModel, [y::FPVector];
                             wts::Union{Nothing, FPVector} = nothing,
                             offset::Union{Nothing, FPVector} = nothing,
                             quantile::Union{Nothing, AbstractFloat} = nothing,
                             ridgeλ::Union{Nothing, Real} = nothing,
                             kwargs...)

Refit the RobustLinearModel.

This function assumes that m was correctly initialized and the model is refitted with the new values for the response, weights, offset, quantile and ridge shrinkage.

Defining a new quantile is only possible for GeneralizedQuantileEstimator.

Defining a new ridgeλ is only possible for RidgePred objects.

refit!(m::QuantileRegression,
      [y::FPVector ;
       verbose::Bool=false,
       quantile::Union{Nothing,
       AbstractFloat}=nothing,
      ]
)

Refit the QuantileRegression model with the new values for the response, weights and quantile. This function assumes that m was correctly initialized.

Model methods

StatsAPI.coef — Function

coef(m::RobustLinearModel)

The coefficients of the model.

Missing docstring.

Missing docstring for StatsBase.coeftable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.confint. Check Documenter's build log for details.

StatsAPI.deviance — Function

deviance(m::RobustLinearModel)

The sum of twice the loss/objective applied to the scaled residuals.

It is consistent with the definition of the deviance for OLS.

deviance(obj::LinearModel)

For linear models, the deviance is equal to the residual sum of squares (RSS).

StatsAPI.nulldeviance — Function

nulldeviance(obj::LinearModel)

For linear models, the deviance of the null model is equal to the total sum of squares (TSS).

Missing docstring.

Missing docstring for StatsBase.dof. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.dof_residual. Check Documenter's build log for details.

StatsAPI.nobs — Method

nobs(obj::RobustResp)::Integer

For linear and generalized linear models, returns the number of elements of the response.

Missing docstring.

Missing docstring for StatsBase.isfitted. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.islinear. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.loglikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.nullloglikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.stderror. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.vcov. Check Documenter's build log for details.

StatsAPI.weights — Function

weights(vs::AbstractArray{<:Real})

Construct a Weights vector from array vs. See the documentation for Weights for more details.

RobustModels.workingweights — Function

workingweights(m::RobustLinearModel)

The robust weights computed by the model.

This can be used to detect outliers, as outliers weights are lower than the weights of valid data points.

Missing docstring.

Missing docstring for StatsBase.fitted. Check Documenter's build log for details.

StatsAPI.predict — Function

predict(mm::LinearModel, newx::AbstractMatrix;
        interval::Union{Symbol,Nothing} = nothing, level::Real = 0.95)

If interval is nothing (the default), return a vector with the predicted values for model mm and new data newx. Otherwise, return a vector with the predicted values, as well as vectors with the lower and upper confidence bounds for a given level (0.95 equates alpha = 0.05). Valid values of interval are :confidence delimiting the uncertainty of the predicted relationship, and :prediction delimiting estimated bounds for new data points.

predict(mm::AbstractGLM, newX::AbstractMatrix; offset::FPVector=eltype(newX)[],
        interval::Union{Symbol,Nothing}=nothing, level::Real = 0.95,
        interval_method::Symbol = :transformation)

Return the predicted response of model mm from covariate values newX and, optionally, an offset.

If interval=:confidence, also return upper and lower bounds for a given coverage level. By default (interval_method = :transformation) the intervals are constructed by applying the inverse link to intervals for the linear predictor. If interval_method = :delta, the intervals are constructed by the delta method, i.e., by linearization of the predicted response around the linear predictor. The :delta method intervals are symmetric around the point estimates, but do not respect natural parameter constraints (e.g., the lower bound for a probability could be negative).

Missing docstring.

Missing docstring for StatsBase.leverage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.modelmatrix. Check Documenter's build log for details.

RobustModels.projectionmatrix — Function

projectionmatrix(m::RobustLinearModel)

The robust projection matrix from the predictor: X (X' W X)⁻¹ X' W

GLM.dispersion — Method

dispersion(m::RobustLinearModel, sqr::Bool=false)

The dispersion is the (weighted) sum of robust residuals. If sqr is true, return the squared dispersion.

Missing docstring.

Missing docstring for StatsBase.response. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StatsBase.residuals. Check Documenter's build log for details.

RobustModels.scale — Function

scale(m::RobustLinearModel, sqr::Bool=false)

The robust scale estimate used for the robust estimation.

If sqr is true, the square of the scale is returned.

RobustModels.tauscale — Function

tauscale(m::RobustLinearModel, sqr::Bool=false; kwargs...)

The robust τ-scale that is minimized in τ-estimation.

If sqr is true, the square of the τ-scale is returned.

RobustModels.location_variance — Function

location_variance(r::RobustLinResp, sqr::Bool=false)

Compute the part of the variance of the coefficients β that is due to the encertainty from the location. If sqr is false, return the standard deviation instead.

From Maronna et al., Robust Statistics: Theory and Methods, Equation 4.49

RobustModels.Estimator — Function

Estimator(m::RobustLinearModel)

The robust estimator object used to fit the model.

Estimators

RobustModels.MEstimator — Type

MEstimator{L<:LossFunction} <: AbstractMEstimator

M-estimator for a given loss function.

The M-estimator is obtained by minimizing the loss function:

\[\hat{\mathbf{\beta}} = \underset{\mathbf{\beta}}{\textrm{argmin}} \sum_{i=1}^n \rho\left(\dfrac{r_i}{\hat{\sigma}}\right)\]

with the residuals $\mathbf{r} = \mathbf{y} - \mathbf{X} \mathbf{\beta}$ , and a robust scale estimate $\hat{\sigma}$.

Fields

loss: the LossFunction used for the robust estimation.

RobustModels.L1Estimator — Type

L1Estimator is a shorthand name for MEstimator{L1Loss}. Using exact QuantileRegression should be prefered.

RobustModels.L2Estimator — Type

L2Estimator is a shorthand name for MEstimator{L2Loss}, the non-robust OLS.

RobustModels.SEstimator — Type

SEstimator{L<:BoundedLossFunction} <: AbstractMEstimator

S-estimator for a given bounded loss function.

The S-estimator is obtained by minimizing the scale estimate:

\[\hat{\mathbf{\beta}} = \underset{\mathbf{\beta}}{\textrm{argmin }} \hat{\sigma}^2\]

where the robust scale estimate $\hat{\sigma}}$ is solution of:

\[\dfrac{1}{n} \sum_{i=1}^n \rho\left(\dfrac{r_i}{\hat{\sigma}}\right) = \delta\]

with the residuals $\mathbf{r} = \mathbf{y} - \mathbf{X} \mathbf{\beta}$ , $\rho$ is a bounded loss function with $\underset{r \to \infty}{\lim} \rho(r) = 1$ and $\delta$ is the finite breakdown point, usually 0.5.

Fields

loss: the LossFunction used for the robust estimation.

RobustModels.MMEstimator — Type

MMEstimator{L1<:BoundedLossFunction, L2<:LossFunction} <: AbstractMEstimator

MM-estimator for the given loss functions.

The MM-estimator is obtained using a two-step process:

compute a robust scale estimate with a high breakdown point using a S-estimate and the loss function L1.

compute an efficient estimate using a M-estimate with the loss function L2.

Fields

loss1: the BoundedLossFunction used for the high breakdown point S-estimation.

loss2: the LossFunction used for the efficient M-estimation.

scaleest: boolean specifying the if the estimation is in the S-estimation step (true)

or the M-estimation step (false).

RobustModels.TauEstimator — Type

TauEstimator{L1<:BoundedLossFunction, L2<:BoundedLossFunction} <: AbstractMEstimator

τ-estimator for the given loss functions.

The τ-estimator corresponds to a M-estimation, where the loss function is a weighted sum of a high breakdown point loss and an efficient loss. The weight is recomputed at every step of the Iteratively Reweighted Least Square, so the estimate is both robust (high breakdown point) and efficient.

Fields

loss1: the high breakdown point BoundedLossFunction.

loss2: the high efficiency BoundedLossFunction.

w: the weight in the sum of losses: w . loss1 + loss2.

RobustModels.GeneralizedQuantileEstimator — Type

GeneralizedQuantileEstimator{L<:LossFunction} <: AbstractQuantileEstimator

Generalized Quantile Estimator is an M-Estimator with asymmetric loss function.

For L1Loss, this corresponds to quantile regression (although it is better to use quantreg for quantile regression because it gives the exact solution).

For L2Loss, this corresponds to Expectile regression (see ExpectileEstimator).

Fields

loss: the LossFunction.

τ: the quantile value to estimate, between 0 and 1.

Properties

tau, q, quantile are aliases for τ.

RobustModels.ExpectileEstimator — Type

The expectile estimator is a generalization of the L2 estimator, for other quantile τ ∈ [0,1].

[1] Schnabel, Eilers - Computational Statistics and Data Analysis 53 (2009) 4168–4177 - Optimal expectile smoothing doi:10.1016/j.csda.2009.05.002

RobustModels.QuantileEstimator — Type

Non-exact quantile estimator, GeneralizedQuantileEstimator{L1Loss}. Prefer using QuantileRegression

Loss functions

RobustModels.L2Loss — Type

The (convex) L2 loss function is that of the standard least squares problem. ψ(r) = r

RobustModels.L1Loss — Type

The standard L1 loss function takes the absolute value of the residual, and is convex but non-smooth. It is not a real L1 loss but a Huber loss with very small tuning constant. ψ(r) = sign(r) Use $QuantileRegression$ for a correct implementation of the L1 loss.

RobustModels.HuberLoss — Type

The convex Huber loss function switches from between quadratic and linear cost/loss function at a certain cutoff. ψ(r) = (abs(r) <= 1) ? r : sign(r)

RobustModels.L1L2Loss — Type

The convex L1-L2 loss interpolates smoothly between L2 behaviour for small residuals and L1 for outliers. ψ(r) = r / √(1 + r^2)

RobustModels.FairLoss — Type

The (convex) "fair" loss switches from between quadratic and linear cost/loss function at a certain cutoff, and is C3 but non-analytic. ψ(r) = r / (1 + abs(r))

RobustModels.LogcoshLoss — Type

The convex Log-Cosh loss function ψ(r) = tanh(r)

RobustModels.ArctanLoss — Type

The convex Arctan loss function ψ(r) = atan(r)

RobustModels.CauchyLoss — Type

The non-convex Cauchy loss function switches from between quadratic behaviour to logarithmic tails. This rejects outliers but may result in multiple minima. For scale estimate, r.ψ(r) is used as a loss, which is the same as for Geman loss. ψ(r) = r / (1 + r^2)

RobustModels.GemanLoss — Type

The non-convex Geman-McClure for strong supression of outliers and does not guarantee a unique solution. For the S-Estimator, it is equivalent to the Cauchy loss. ψ(r) = r / (1 + r^2)^2

RobustModels.WelschLoss — Type

The non-convex Welsch for strong supression of outliers and does not guarantee a unique solution ψ(r) = r * exp(-r^2)

RobustModels.TukeyLoss — Type

The non-convex Tukey biweight estimator which completely suppresses the outliers, and does not guaranty a unique solution. ψ(r) = (abs(r) <= 1) ? r * (1 - r^2)^2 : 0

RobustModels.YohaiZamarLoss — Type

The non-convex (and bounded) optimal Yohai-Zamar loss function that minimizes the estimator bias. It was originally introduced in Optimal locally robust M-estimates of regression (1997) by Yohai and Zamar with a slightly different formula.

Estimator and Loss functions methods

RobustModels.rho — Function

The loss function ρ for the M-estimator.

RobustModels.psi — Function

The influence function ψ is the derivative of the loss function for the M-estimator, multiplied by the square of the tuning constant.

RobustModels.psider — Function

The derivative of ψ, used for asymptotic estimates.

RobustModels.weight — Function

The weights for IRLS, the function ψ divided by r.

RobustModels.values — Function

Faster version if you need ρ, ψ and w in the same call

RobustModels.estimator_norm — Function

The integral of exp(-ρ) used for calculating the full-loglikelihood

RobustModels.estimator_bound — Function

The limit at ∞ of the loss function. Used for scale estimation of bounded loss.

RobustModels.tuning_constant — Function

The tuning constant of the loss function, can be optimized to get efficient or robust estimates.

RobustModels.isconvex — Function

Boolean if the estimator or loss function is convex

RobustModels.isbounded — Function

Boolean if the estimator or loss function is bounded

RobustModels.estimator_high_breakdown_point_constant — Function

The tuning constant associated to the loss that gives an efficient M-estimator.

RobustModels.estimator_high_efficiency_constant — Function

The tuning constant associated to the loss that gives a robust (high breakdown point) M-estimator.

RobustModels.efficient_loss — Function

The loss initialized with an efficient tuning constant

RobustModels.robust_loss — Function

The loss initialized with a robust (high breakdown point) tuning constant

RobustModels.efficiency_tuning_constant — Function

The tuning constant c is computed so the efficiency for Normal distributed residuals is 0.95. The efficiency of the mean estimate μ is defined by:

eff_μ = (E[ψ'])²/E[ψ²]

RobustModels.mscale_loss — Function

mscale_loss(loss::L, x)

The rho-function that is used for M-scale estimation.

For monotone (convex) functions, χ(r) = r.ψ(r).

For bounded functions, χ(r) = ρ(r)/ρ(∞) so χ(∞) = 1.

RobustModels.breakdown_point_tuning_constant — Function

The M-estimate of scale is computed by solving:

\[\dfrac{1}{n} \sum_i \chi\left( \dfrac{r_i}{\hat{\sigma}}\right) = \delta\]

For monotone (convex) functions, χ(r) = r.ψ(r) and δ is defined as E[χ(r)] = δ for the Normal distribution N(0,1) For bounded functions, χ(r) = ρ(r)/ρ(∞) with χ(∞) = 1 and δ = E[χ]/χ(∞) with expectation w.r.t. Normal density.

The tuning constant c corresponding to a high breakdown point (0.5) is such that δ = 1/2, from 1/n Σ χ(r/ŝ) = δ

RobustModels.scale_estimate — Function

scale_estimate(loss, res; σ0=1.0, wts=[], verbose=false,
                         order=1, approx=false, nmax=30,
                         rtol=1e-4, atol=0.1)

Compute the M-scale estimate from the loss function. If the loss is bounded, ρ is used as the function χ in the sum, otherwise r.ψ(r) is used if the loss is not bounded, to coincide with the Maximum Likelihood Estimator. Also, for bounded estimator, because f(s) = 1/(nδ) Σ ρ(ri/s) is decreasing the iteration step is not using the weights but is multiplicative.

RobustModels.tau_efficiency_tuning_constant — Function

tau_efficiency_tuning_constant(::Type{L1}, ::Type{L2}; eff::Real=0.95, c0::Real=1.0)
    where {L1<:BoundedLossFunction,L2<:BoundedLossFunction}

Compute the tuning constant that corresponds to a high breakdown point for the τ-estimator.

RobustModels.estimator_tau_efficient_constant — Function

The tuning constant associated to the loss that gives a robust τ-estimator.

RobustModels.loss — Function

The loss function used for the estimation

RobustModels.set_SEstimator — Function

MEstimator, set to S-Estimation phase

RobustModels.set_MEstimator — Function

MEstimator, set to M-Estimation phase

RobustModels.update_weight! — Function

update_weight!(E::TauEstimator, res::AbstractArray{T}; wts::AbstractArray{T}=T[])

Update the weight between the two estimators of a τ-estimator using the scaled residual.

RobustModels.tau_scale_estimate — Function

tau_scale_estimate!(E::TauEstimator, res::AbstractArray{T}, σ::Real, sqr::Bool=false;
                    wts::AbstractArray{T}=T[], bound::AbstractFloat=0.5) where {T<:AbstractFloat}

The τ-scale estimate, where σ is the scale estimate from the robust M-scale. If sqr is true, return the squared value.

RobustModels.quantile_weight — Function

quantile_weight(τ::Real, r::Real)

Wrapper function to compute quantile-like loss function.